+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.16.9] - 2021-09-26
+
+### Added
+
+-   Added passwordProtected fallback value for the status embed
+
+### Fixed
+
+-   Fixed server name fallback not being used properly (also shows the password-protected status fallback)
+
 ## [1.16.8] - 2021-09-26
 
 ### Added